cmd/snap-update-ns: move applyProfile to utils.go
